+++ shortcut.rkt
@@ -1,6 +1,7 @@
 #lang s-exp framework/keybinding-lang
 
-(require "panel/version-control.rkt"
+(require "meta.rkt"
+         "panel/version-control.rkt"
          "project-manager.rkt")
 
 (define (c+ k)
@@ -105,3 +106,17 @@
                    [on-select
                     (λ (path)
                       (current-project path))])))
+
+(define (replace-special-symbol editor event)
+  (let* ([end (send editor get-start-position)]
+         [start (send editor get-backward-sexp end)]
+         [to-complete (send editor get-text start end)])
+    (when (string-prefix? to-complete "\\")
+      ;;; select previous sexp
+      (send editor set-position start end)
+      ;;; replace it with new text
+      (send editor insert (hash-ref latex-complete (string-trim to-complete "\\" #:right? #f)
+                                    to-complete))))
+  (send editor insert (send event get-key-code)))
+(keybinding "space" replace-special-symbol)
+(keybinding "enter" replace-special-symbol)